refactor: App의 공통된 로직들 Domain 폴더로 이동 및 App을 WebApp으로 파일 이름 변경
